Remove old component and add window listner
